requirements.txt
streamlit
pytesseract
pillow
opencv-python
plotly
pandas
requests
groq
python-dotenv
